Replace optional transition times with fixed time
This marks a change to the way transition data was calculated and stored.

Previously, each timespan—a period of consistent time zone name, GMT offset, and DST offset—was stored as a transition with an optional timestamp at which the transition occurs. This doesn’t solve the problem of what the time details should be *before* the first transition has occurred, which is why the timestamp was optional: the first period would simply have a `None` timestamp.

This approach had problems, namely:

1. It was possible to have multiple `None` timestamps, which would confuse the datetime calculator;
2. It was not only possible but actually manifest to have time zones with *zero* `None` timestamps, which would also confuse the datetime calculator.

To remove the possibility of having invalid data, each time zone now has one fixed “ante” period, and at least zero transitions, each of which now has a non-Optional timestamp.

It’s a rather simple change, but most of the terminology has been updated to reflect this—Transition to ZoneDetails, among other structs—which is where most of the lines in the diff come from.

The only time zones that change in zoneinfo-data are the “fixed” ones like CET or EST5EDT.