Begin refactoring the builder code into methods
This moves most of the 'convert this time zone into a list of timespans' code to a new Builder type.

The refactoring isn't yet complete, hopefully obviously, but it's at a stable point when it compiles and the data crate output is not modified.

It highlights how many of the variables being changed on each iteration are only important for one or the other Saving variant. It also combines the code for the NoSaving and Fixed variants, as they differed only by one variable.

Next steps include fixing some of the warnings and getting rid of more per-iteration variables.