Move filter-by-regex into a crate so that we can use cargo features
This makes it possible to only install the regex build dependency with the
`features = ["filter-by-regex"]` option in Cargo.toml.

This commit is broken -- build.rs will get a new trivial implementation in the
next commit. Deleting and creating the file in separate commits causes git's
rename logic to kick in, making the diff much more useful.