use a perfect hash table to parse timezones
This commit also enables the new cargo resolver. This is necessary
because the build time dependency `chrono-tz-build` depends on
`phf_shared` with its `std` feature enabled, but that causes the
chrono-tz library as a whole to fail to compile on `no_std` targets.

The new resolver correctly separates crate features of build and dev
dependencies from those of target dependencies as explain in the release
notes:

https://blog.rust-lang.org/2021/03/25/Rust-1.51.0.html#cargos-new-feature-resolver

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>